fix(AppPage Files): Change location of file modules
