+++ src/Pages/AppPage/Utilities/GridUtilities.ts
@@ -1,5 +1,3 @@
-import Position from '../Types/Position';
-
 export const createGrid = (size: number): boolean[][] => {
   const grid: boolean[][] = Array<boolean[]>();
 
@@ -35,29 +33,4 @@ export const changeGridCell = (
   newGrid[cellY][cellX] = value;
 
   return newGrid;
-};
-
-export const paintGrid = (
-  grid: boolean[][],
-  ctx: CanvasRenderingContext2D,
-  canvasResolution: number,
-  cellSize: number,
-  position: Position
-): void => {
-  ctx.fillRect(0, 0, canvasResolution, canvasResolution);
-  const cellsInRes = canvasResolution / cellSize;
-
-  for (let row = position.x; row < cellsInRes && row < grid.length; row++) {
-    for (
-      let col = position.y;
-      col < cellsInRes && col < grid[row].length;
-      col++
-    ) {
-      const startX = (col - position.x) * cellSize;
-      const startY = (row - position.y) * cellSize;
-
-      ctx.fillStyle = grid[row][col] ? 'white' : 'black';
-      ctx.fillRect(startX, startY, cellSize, cellSize);
-    }
-  }
 };
